backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Nikita Glazkov <zitrnik@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.2.2"
psycopg2-binary = "^2.9.5"
gunicorn = "^20.1.0"
flask-wtf = "^1.0.1"
flask-login = "^0.6.2"
wtforms-components = "^0.10.5"
flask-bootstrap = "^3.3.7.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
